initial react router setup and refactoring to display buttons which link to correct pages
