+++ src/components/buttons/CreateItemButton.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faPlus } from '@fortawesome/free-solid-svg-icons'
-import "../styles/navbar-buttons.css"
 
 
 const CreateItemButton = () => {
