+++ src/components/buttons/HomeButton.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faHouse } from '@fortawesome/free-solid-svg-icons'
-import "../styles/navbar-buttons.css"
+
 
 
 const HomeButton = () => {
