+++ src/components/buttons/LogoutButton.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faRightFromBracket } from '@fortawesome/free-solid-svg-icons'
-import "../styles/navbar-buttons.css"
+
 
 const LogoutButton = () => {
   return (
