+++ src/components/pages/CreateItemPage.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const CreateItemPage = () => {
+    return(
+        <h1>Add a new Shirt</h1>
+    )
+}
+
+export default CreateItemPage
