+++ src/components/pages/LogoutPage.js
@@ -1,0 +1,10 @@
+import React from 'react'
+
+const LogoutPage = () => {
+  return (
+    <h1>Logout Page</h1>
+    )
+
+}
+
+export default LogoutPage
