+++ src/components/pages/ProfilePage.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const ProfilePage = () => {
+    return(
+    <h1>My Profile</h1>
+    )
+}
+
+export default ProfilePage
